Isolate UI dependencies with feature flags
