+++ crates/rmf_site_sim/src/interaction/mod.rs
@@ -1,0 +1,7 @@
+pub mod keyboard;
+
+#[cfg(feature = "egui")]
+pub mod egui;
+
+#[cfg(feature = "rmf_site_egui")]
+pub mod rmf_site_egui;
